requirements.txt
# Django
# psycopg2
# gunicorn
asgiref==3.5.2
attrs==21.4.0
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.5.1
PyJWT==1.7.1
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
psycopg==3.0.14
gunicorn==20.1.0
django-tabbed-admin==1.0.4
django-import-export==2.7.1
django-widget-tweaks==1.4.9
django-reversion==4.0.1